Add function to set prescribed values at dofs
The `solve` function is made slightly more generic by allowing to pass
an iteratable collection containing pairs of dofs and their value. The
selction of dofs and their values can be generated by the helper routine
prescribe, which filters relevant nodes given a predicate function
operating on the nodal coordinate and a separate mapping function to map
the dofs coordinate to the set value.

Along the way some helper functions were introduced to allow iteration
over nodes and coordinates present in a mesh. At relevant places these
helpers are employed.