Evaluate shape functions at Element construction
The patch replaces the hardcoded shape function values for the one
dimensional element with shape functions that are generated as the
cartesian product of the one-dimensional element.

The Element struct is updated to reflect that N is expected to be any
AbstractMatrix. In practice, this is a SMatrix, given that the size and
values are known at compile time and will not change throughout.