Add struct for quadrature information
Still unsure if more than one quadrature point make sense for this
kind of implementation, as that would require reevaluating/rebuilding
N,B,K within the assemble/integration loops. This is currently not done
and would require some additional computational cost that might not be
worth it?

Regardless, this patch includes a small struct to contain the quadrature
information and make it somewhat more explicit that the previous version
with just a hardcoded weight and implicit locations. At least the values
are directly picked up from the quadrature struct in setting up the
element struct.