Extend Mesh with support for subset meshes
The Mesh struct is extended to carry the elements, nodes, and node
offset iterators directly, where the elements and nodes iterators are
extended with support for setting the origin index as well as their
increment.

This allows the Mesh struct to describe boundary meshes too. For those
meshes, the origin might not align with the origin of the original mesh,
and the elements/nodes iterators do not necessarily increment with unit
steps, for example when the boundary is aligned with the y, or z-axis.