Decouple mesh and element dimensions
The mesh dimension Mesh{Dim} dictates the global dimensions in which the
mesh is embedded and controls the dimensionality of elements, nodes, and
coordinates.

Within this coordinate system we can now have meshes and boundary meshes
that are defined within the same dimensionality, but might have elements
of lower dimension, e.g. the boundary of a two-dimensional mesh will
have one-dimensional elements.

This distinction was not captured in the implementation as we only
looked at the dimension of the mesh for any questions on dimensionality.
Now, we have the need to look at the actual element dimension that is
embedded within the dimensionality of the mesh.

The embedded dimension can be extracted from the mesh's offsets and
inferred from the count of node offsets, which triggers some further
refactoring of the assembly and integration routines to decouple some of
the parameters from the Element and calculate those on the fly.