wip: Provide plot functions by Julia extension
TODO: Still to do the actual implementation, this only illustrates the
concept using the dummy function.

The plot functionality will be provided as extension that is only loaded
when GLMakie and LaTeXStrings are loaded in the environment before MinX
is loaded, such that the plot related packages only need to be weak
dependencies and are not required for the basic use of MinX.

It seems that functions can only be "extended", so dummy interfaces are
added withitn MinX that are extended with the actual interfaces in the
extension only.